refactor: PassengerSummary type 패키지로 이동 (#164)
