+++ src/main/java/com/sudo/railo/booking/application/ReservationService.java
@@ -19,9 +19,9 @@ import com.sudo.railo.booking.application.dto.request.ReservationDeleteRequest;
 import com.sudo.railo.booking.application.dto.response.ReservationDetail;
 import com.sudo.railo.booking.application.dto.response.SeatReservationDetail;
 import com.sudo.railo.booking.config.BookingConfig;
-import com.sudo.railo.booking.domain.PassengerSummary;
 import com.sudo.railo.booking.domain.Reservation;
 import com.sudo.railo.booking.domain.status.ReservationStatus;
+import com.sudo.railo.booking.domain.type.PassengerSummary;
 import com.sudo.railo.booking.exception.BookingError;
 import com.sudo.railo.booking.infrastructure.reservation.ReservationRepository;
 import com.sudo.railo.booking.infrastructure.reservation.ReservationRepositoryCustom;
