+++ src/main/java/com/sudo/railo/booking/application/dto/request/ReservationCreateRequest.java
@@ -2,7 +2,7 @@ package com.sudo.railo.booking.application.dto.request;
 
 import java.util.List;
 
-import com.sudo.railo.booking.domain.PassengerSummary;
+import com.sudo.railo.booking.domain.type.PassengerSummary;
 import com.sudo.railo.booking.domain.type.TripType;
 
 import io.swagger.v3.oas.annotations.media.Schema;
