+++ src/main/java/com/sudo/railo/booking/domain/type/PassengerSummary.java
@@ -1,6 +1,4 @@
-package com.sudo.railo.booking.domain;
-
-import com.sudo.railo.booking.domain.type.PassengerType;
+package com.sudo.railo.booking.domain.type;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
